self_replication_service__mycelium/mycelium-simulation/requirements.txt
pyyaml>=6.0
matplotlib>=3.8
